@@ -11,16 +11,16 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "^17.3.12",
+    "@angular/animations": "^18.1.3",
     "@angular/cdk": "^18.1.3",
-    "@angular/common": "^17.3.12",
-    "@angular/compiler": "^17.3.12",
-    "@angular/core": "^17.3.12",
-    "@angular/forms": "^17.3.12",
+    "@angular/common": "^18.1.3",
+    "@angular/compiler": "^18.1.3",
+    "@angular/core": "^18.1.3",
+    "@angular/forms": "^18.1.3",
     "@angular/material": "^18.1.3",
-    "@angular/platform-browser": "^17.3.12",
-    "@angular/platform-browser-dynamic": "^17.3.12",
-    "@angular/router": "^17.3.12",
+    "@angular/platform-browser": "^18.1.3",
+    "@angular/platform-browser-dynamic": "^18.1.3",
+    "@angular/router": "^18.1.3",
     "@highcharts/map-collection": "^2.2.0",
     "apexcharts": "^3.51.0",
     "highcharts": "^11.4.6",
@@ -34,7 +34,7 @@
   "devDependencies": {
     "@angular-devkit/build-angular": "^17.3.8",
     "@angular/cli": "~17.3.8",
-    "@angular/compiler-cli": "^17.3.12",
+    "@angular/compiler-cli": "^18.1.3",
     "@types/jasmine": "~4.3.0",
     "autoprefixer": "^10.4.19",
     "jasmine-core": "~4.6.0",
